cutting stuff

+++ quake_framebuffer/quake_framebuffer/d_fill.cpp
@@ -48,9 +48,9 @@ void D_FillRect (vrect_t *rect, int color)
 		rheight += ry;
 		ry = 0;
 	}
-	if (rx+rwidth > vid.width)
+	if (static_cast<size_t>(rx+rwidth) > vid.width)
 		rwidth = vid.width - rx;
-	if (ry+rheight > vid.height)
+	if (static_cast<size_t>(ry+rheight) > vid.height)
 		rheight = vid.height - rx;
 		
 	if (rwidth < 1 || rheight < 1)
